fix(lint): resolve 145 ruff 0.16 errors to unblock dependabot PRs
ruff 0.16 introduced new rules (FURB162, UP045, UP017, I001, etc.) that
broke all dependabot CI runs. Auto-fixed 115 via --fix (modern typing
Optional→X|None, datetime.UTC shorthand, import sorting). Remaining 30
fixed manually: real code improvements (use .items() instead of dict
index, remove redundant int() cast, add check=True to subprocess.run,
combine nested with statements, use next() over list slice) and
targeted noqa for intentional patterns (BLE001 broad-catch in providers
for resilience, DTZ011/DTZ007/DTZ001 for date-only logic). Bumped ruff
floor from 0.15.20 to 0.16.0 so CI and local stay in sync. All 370
backend tests pass.

+++ backend/src/scoring/mapping.py
@@ -1,5 +1,6 @@
 """映射分 — 美股 vs A 股 ETF 的对齐 60d 相关性 × 100"""
 import math
+
 import pandas as pd  # type: ignore[import-untyped]
 from scipy.stats import pearsonr  # type: ignore[import-untyped]
 
@@ -46,4 +47,4 @@ def mapping_score(
     corr, _ = pearsonr(aligned['us'], aligned['cn'])
     if math.isnan(corr):
         return None
-    return int(round(abs(corr) * 100))
+    return round(abs(corr) * 100)
